functions/rss/requirements.txt
# Core functionality
firebase-admin==6.6.0
functions-framework==3.9.2

# RSS parsing
feedparser==6.0.11

# Utilities
python-dotenv==1.0.1
